moved find_datasets, find_references to catalog
